Add yield support to manager high interface
Adding a SessionRunProgress message to hold the progress of
a run call, it contains an application_progress based on the
yield and a progress based on the Cartesi Machine cycle. It
also holds the a timestamp of when the progress was updated
and the last known cycle of the Cartesi Machine.

Also changing all manager high methods response naming convention from
Session<METHOD>Result to Session<METHOD>Response to be consistent with
the other interfaces.